Workaround to be able to fix https://github.com/sass/node-sass/issues/1904
To be able to install a `node-sass` version higher than 3.13.1 (which reports a 404), so that it is not necessary compiling `libsass`, an upgrade from `gulp-sass` is needed. After that, the `node-sass` version contains the referenced bug: "*index variable reference issues in for loops within functions*". The bug has been fixed, yet for a `libsass` version 3.5.0.beta.3, which is not yet used by any other `node-sass` version.

The workaround basically consists on  creating a dummy mathematical operation when referencing the index variable of the loop. It seems like this, that the index is assigned as a new memory variable, instead of as a reference.